// src/actions/action-keys.enum.ts
export enum ActionKey {
  BASIC_WISH_FOR_CHILDREN = 'basic-wish-for-children',
  CATTLE_MARKET = 'cattle-market',
  CULTIVATION = 'cultivation',
  EASTERN_QUARRY = 'eastern-quarry',
  FARM_REDEVELOPMENT = 'farm-redevelopment',
  FENCING = 'fencing',
  GRAIN_UTILIZATION = 'grain-utilization',
  HOUSE_REDEVELOPMENT = 'house-redevelopment',
  MAJOR_IMPROVEMENT = 'major-improvement',
  PIG_MARKET = 'pig-market',
  SHEEP_MARKET = 'sheep-market',
  URGENT_WISH_FOR_CHILDREN = 'urgent-wish-for-children',
  VEGETABLE_SEEDS = 'vegetable-seeds',
  WESTERN_QUARRY = 'western-quarry',
  CLAY_PIT = 'clay-pit',
  DAY_LABORER = 'day-laborer',
  FARM_EXPANSION = 'farm-expansion',
  FARMLAND = 'farmland',
  FISHING = 'fishing',
  FOREST = 'forest',
  GRAIN_SEEDS = 'grain-seeds',
  LESSONS = 'lessons',
  MEETING_PLACE = 'meeting-place',
  REED_BANK = 'reed-bank',
}